Se separan tests de codigo
