+++ TSE_TP4/get_supercap_voltage.py
@@ -1,4 +1,3 @@
-import pytest
 import time
 
 VALOR_CRITICO = 1965
@@ -14,7 +13,7 @@ def sleep_ms(sleep_time_in_miliseconds):
     time.sleep(sleep_time_in_miliseconds/1000)
 
 #Attach the method "sleep_ms" to the existingclass on runtim
-setattr(time, "sleep_ms", sleep_ms) 
+setattr(time, "sleep_ms", sleep_ms)
 
 class adc:
     """
@@ -66,48 +65,4 @@ def get_ADC_value(adc_input_pin):
         print("The chosen pin is not a valid ADC port or not a string with the format 'Pxx'")
         print("Received ", adc_input_pin)
         return -3
-
-###-------------------------TESTS--------------------------------###
-def test_NO_habilitar_3v3_si_vcap_es_menor(mocker):
-    mocker.patch(__name__ + ".get_ADC_value", return_value=1964)
-    switch_3v3 = enable_3v3()
-    assert switch_3v3 == False
-
-def test_SI_habilitar_3v3_si_vcap_es_igual(mocker):
-    mocker.patch(__name__ + ".get_ADC_value", return_value=1965)
-    switch_3v3 = enable_3v3()
-    assert switch_3v3 == True
-
-def test_SI_habilitar_3v3_si_vcap_es_mayor(mocker):
-    mocker.patch(__name__ + ".get_ADC_value", return_value=1966)
-    switch_3v3 = enable_3v3()
-    assert switch_3v3 == True
-
-def test_NO_habilitar_5V_si_vcap_es_menor(mocker):
-    mocker.patch(__name__ + ".get_ADC_value", return_value=1964)
-    switch_5V = enable_5V()
-    assert switch_5V == False
-
-def test_SI_habilitar_5V_si_vcap_es_igual(mocker):
-    mocker.patch(__name__ + ".get_ADC_value", return_value=1965)
-    switch_5V = enable_5V()
-    assert switch_5V == True
-
-def test_SI_habilitar_5V_si_vcap_es_mayor(mocker):
-    mocker.patch(__name__ + ".get_ADC_value", return_value=1966)
-    switch_5V = enable_5V()
-    assert switch_5V == True
-
-def test_get_ADC_value_invalid_pin_integer():
-    value = get_ADC_value(13)
-    assert value == -3
-
-def test_get_ADC_value_from_invalid_pin_non_analog():
-    value = get_ADC_value('P12')
-    assert value == -3
-
-def test_get_ADC_from_valid_pin():
-    value = get_ADC_value('P13')
-    assert value == 1965
-    assert isinstance(value, int)
